% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromeosdevices.R
\name{chromeosdevices.get}
\alias{chromeosdevices.get}
\title{Retrieves a Chrome OS device's properties.}
\usage{
chromeosdevices.get(
  customerId,
  deviceId,
  projection = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.chromeosdevices.get"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customerId}{The unique ID for the customer's G Suite account. As an account administrator, you can also use the \code{my_customer} alias to represent your account's \code{customerId}. The \code{customerId} is also returned as part of the \href{/admin-sdk/directory/v1/reference/users}{Users resource}.}

\item{deviceId}{The unique ID of the device. The \code{deviceId}s are returned in the response from the \href{/admin-sdk/directory/v1/reference/chromeosdevices/list}{chromeosdevices.list} method.}

\item{projection}{Determines whether the response contains the full list of properties or only a subset.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.device.chromeos
\item https://www.googleapis.com/auth/admin.directory.device.chromeos.readonly
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
