% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromeosdevices.R
\name{chromeosdevices.moveDevicesToOu}
\alias{chromeosdevices.moveDevicesToOu}
\title{Move or insert multiple Chrome OS devices to an organizational unit. You can move up to 50 devices at once.}
\usage{
chromeosdevices.moveDevicesToOu(
  customerId,
  orgUnitPath,
  deviceIds = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.chromeosdevices.moveDevicesToOu"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customerId}{Immutable ID of the G Suite account}

\item{orgUnitPath}{Full path of the target organizational unit or its ID}

\item{deviceIds}{Chrome OS devices to be moved to OU}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.device.chromeos
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
