% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainAliases.R
\name{domainAliases.get}
\alias{domainAliases.get}
\title{Retrieves a domain alias of the customer.}
\usage{
domainAliases.get(
  domainAliasName,
  customer,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.domainAliases.get"]]$scopes),
  return_response = F
)
}
\arguments{
\item{domainAliasName}{Name of domain alias to be retrieved.}

\item{customer}{Immutable ID of the G Suite account.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.domain
\item https://www.googleapis.com/auth/admin.directory.domain.readonly
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
