% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainAliases.R
\name{domainAliases.insert}
\alias{domainAliases.insert}
\title{Inserts a domain alias of the customer.}
\usage{
domainAliases.insert(
  customer,
  parentDomainName = NULL,
  kind = NULL,
  domainAliasName = NULL,
  verified = NULL,
  etag = NULL,
  creationTime = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.domainAliases.insert"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customer}{Immutable ID of the G Suite account.}

\item{parentDomainName}{The parent domain name that the domain alias is associated with. This can either be a primary or secondary domain name within a customer.}

\item{kind}{Kind of resource this is.}

\item{domainAliasName}{The domain alias name.}

\item{verified}{Indicates the verification state of a domain alias. (Read-only)}

\item{etag}{ETag of the resource.}

\item{creationTime}{The creation time of the domain alias. (Read-only).}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.domain
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
