% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\name{domains.insert}
\alias{domains.insert}
\title{Inserts a domain of the customer.}
\usage{
domains.insert(
  customer,
  isPrimary = NULL,
  kind = NULL,
  verified = NULL,
  etag = NULL,
  creationTime = NULL,
  domainAliases = NULL,
  domainName = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.domains.insert"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customer}{Immutable ID of the G Suite account.}

\item{isPrimary}{Indicates if the domain is a primary domain (Read-only).}

\item{kind}{Kind of resource this is.}

\item{verified}{Indicates the verification state of a domain. (Read-only).}

\item{etag}{ETag of the resource.}

\item{creationTime}{Creation time of the domain. Expressed in \href{http://en.wikipedia.org/wiki/Epoch_time}{Unix time} format. (Read-only).}

\item{domainAliases}{List of domain alias objects. (Read-only)}

\item{domainName}{The domain name of the customer.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.domain
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
