% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups.delete}
\alias{groups.delete}
\title{Deletes a group.}
\usage{
groups.delete(
  groupKey,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.groups.delete"]]$scopes),
  return_response = F
)
}
\arguments{
\item{groupKey}{Identifies the group in the API request. The value can be the group's email address, group alias, or the unique group ID.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.group
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
