% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{members.list}
\alias{members.list}
\title{Retrieves a paginated list of all members in a group.}
\usage{
members.list(
  groupKey,
  includeDerivedMembership = NULL,
  maxResults = NULL,
  pageToken = NULL,
  roles = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.members.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{groupKey}{Identifies the group in the API request. The value can be the group's email address, group alias, or the unique group ID.}

\item{includeDerivedMembership}{Whether to list indirect memberships. Default: false.}

\item{maxResults}{Maximum number of results to return. Max allowed value is 200.}

\item{pageToken}{Token to specify next page in the list.}

\item{roles}{The \code{roles} query parameter allows you to retrieve group members by role. Allowed values are \code{OWNER}, \code{MANAGER}, and \code{MEMBER}.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.group
\item https://www.googleapis.com/auth/admin.directory.group.member
\item https://www.googleapis.com/auth/admin.directory.group.member.readonly
\item https://www.googleapis.com/auth/admin.directory.group.readonly
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
