% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledevices.R
\name{mobiledevices.action}
\alias{mobiledevices.action}
\title{Takes an action that affects a mobile device. For example, remotely wiping a device.}
\usage{
mobiledevices.action(
  customerId,
  resourceId,
  action = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.mobiledevices.action"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customerId}{The unique ID for the customer's G Suite account. As an account administrator, you can also use the \code{my_customer} alias to represent your account's \code{customerId}. The \code{customerId} is also returned as part of the \href{/admin-sdk/directory/v1/reference/users}{Users resource}.}

\item{resourceId}{The unique ID the API service uses to identify the mobile device.}

\item{action}{The action to be performed on the device.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.device.mobile
\item https://www.googleapis.com/auth/admin.directory.device.mobile.action
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
