% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledevices.R
\name{mobiledevices.list}
\alias{mobiledevices.list}
\title{Retrieves a paginated list of all mobile devices for an account.}
\usage{
mobiledevices.list(
  customerId,
  orderBy = NULL,
  sortOrder = NULL,
  projection = NULL,
  maxResults = NULL,
  pageToken = NULL,
  query = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.mobiledevices.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customerId}{The unique ID for the customer's G Suite account. As an account administrator, you can also use the \code{my_customer} alias to represent your account's \code{customerId}. The \code{customerId} is also returned as part of the \href{/admin-sdk/directory/v1/reference/users}{Users resource}.}

\item{orderBy}{Device property to use for sorting results.}

\item{sortOrder}{Whether to return results in ascending or descending order. Must be used with the \code{orderBy} parameter.}

\item{projection}{Restrict information returned to a set of selected fields.}

\item{maxResults}{Maximum number of results to return. Max allowed value is 100.}

\item{pageToken}{Token to specify next page in the list}

\item{query}{Search string in the format given at https://developers.google.com/admin-sdk/directory/v1/search-operators}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.device.mobile
\item https://www.googleapis.com/auth/admin.directory.device.mobile.action
\item https://www.googleapis.com/auth/admin.directory.device.mobile.readonly
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
