% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgunits.R
\name{orgunits.list}
\alias{orgunits.list}
\title{Retrieves a list of all organizational units for an account.}
\usage{
orgunits.list(
  customerId,
  orgUnitPath = NULL,
  type = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.orgunits.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customerId}{The unique ID for the customer's G Suite account. As an account administrator, you can also use the \code{my_customer} alias to represent your account's \code{customerId}. The \code{customerId} is also returned as part of the \href{/admin-sdk/directory/v1/reference/users}{Users resource}.}

\item{orgUnitPath}{The full path to the organizational unit or its unique ID. Returns the children of the specified organizational unit.}

\item{type}{Whether to return all sub-organizations or just immediate children.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.orgunit
\item https://www.googleapis.com/auth/admin.directory.orgunit.readonly
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
