% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgunits.R
\name{orgunits.update}
\alias{orgunits.update}
\title{Updates an organizational unit.}
\usage{
orgunits.update(
  orgUnitPath,
  customerId,
  etag = NULL,
  blockInheritance = NULL,
  orgUnitId = NULL,
  name = NULL,
  parentOrgUnitPath = NULL,
  kind = NULL,
  parentOrgUnitId = NULL,
  description = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.orgunits.update"]]$scopes),
  return_response = F
)
}
\arguments{
\item{orgUnitPath}{The full path of the organizational unit or its unique ID.}

\item{customerId}{The unique ID for the customer's G Suite account. As an account administrator, you can also use the \code{my_customer} alias to represent your account's \code{customerId}. The \code{customerId} is also returned as part of the \href{/admin-sdk/directory/v1/reference/users}{Users resource}.}

\item{etag}{ETag of the resource.}

\item{blockInheritance}{Determines if a sub-organizational unit can inherit the settings of the parent organization. The default value is \code{false}, meaning a sub-organizational unit inherits the settings of the nearest parent organizational unit. For more information on inheritance and users in an organization structure, see the \href{http://support.google.com/a/bin/answer.py?answer=182442&topic=1227584&ctx=topic}{administration help center}.}

\item{orgUnitId}{The unique ID of the organizational unit.}

\item{name}{The organizational unit's path name. For example, an organizational unit's name within the /corp/support/sales_support parent path is sales_support. Required.}

\item{parentOrgUnitPath}{The organizational unit's parent path. For example, /corp/sales is the parent path for /corp/sales/sales_support organizational unit. Required, unless \code{parentOrgUnitId} is set.}

\item{kind}{The type of the API resource. For Orgunits resources, the value is \code{admin#directory#orgUnit}.}

\item{parentOrgUnitId}{The unique ID of the parent organizational unit. Required, unless \code{parentOrgUnitPath} is set.}

\item{description}{Description of the organizational unit.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.orgunit
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
