% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleAssignments.R
\name{roleAssignments.insert}
\alias{roleAssignments.insert}
\title{Creates a role assignment.}
\usage{
roleAssignments.insert(
  customer,
  assignedTo = NULL,
  kind = NULL,
  etag = NULL,
  roleAssignmentId = NULL,
  scopeType = NULL,
  orgUnitId = NULL,
  roleId = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.roleAssignments.insert"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customer}{Immutable ID of the G Suite account.}

\item{assignedTo}{The unique ID of the user this role is assigned to.}

\item{kind}{The type of the API resource. This is always \code{admin#directory#roleAssignment}.}

\item{etag}{ETag of the resource.}

\item{roleAssignmentId}{ID of this roleAssignment.}

\item{scopeType}{The scope in which this role is assigned.}

\item{orgUnitId}{If the role is restricted to an organization unit, this contains the ID for the organization unit the exercise of this role is restricted to.}

\item{roleId}{The ID of the role that is assigned.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.rolemanagement
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
