% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleAssignments.R
\name{roleAssignments.list}
\alias{roleAssignments.list}
\title{Retrieves a paginated list of all roleAssignments.}
\usage{
roleAssignments.list(
  customer,
  userKey = NULL,
  roleId = NULL,
  maxResults = NULL,
  pageToken = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.roleAssignments.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customer}{Immutable ID of the G Suite account.}

\item{userKey}{The user's primary email address, alias email address, or unique user ID. If included in the request, returns role assignments only for this user.}

\item{roleId}{Immutable ID of a role. If included in the request, returns only role assignments containing this role ID.}

\item{maxResults}{Maximum number of results to return.}

\item{pageToken}{Token to specify the next page in the list.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.rolemanagement
\item https://www.googleapis.com/auth/admin.directory.rolemanagement.readonly
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
