% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{roles.delete}
\alias{roles.delete}
\title{Deletes a role.}
\usage{
roles.delete(
  roleId,
  customer,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.roles.delete"]]$scopes),
  return_response = F
)
}
\arguments{
\item{roleId}{Immutable ID of the role.}

\item{customer}{Immutable ID of the G Suite account.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.rolemanagement
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
