% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{roles.list}
\alias{roles.list}
\title{Retrieves a paginated list of all the roles in a domain.}
\usage{
roles.list(
  customer,
  maxResults = NULL,
  pageToken = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.roles.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customer}{Immutable ID of the G Suite account.}

\item{maxResults}{Maximum number of results to return.}

\item{pageToken}{Token to specify the next page in the list.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.rolemanagement
\item https://www.googleapis.com/auth/admin.directory.rolemanagement.readonly
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
