% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{roles.update}
\alias{roles.update}
\title{Updates a role.}
\usage{
roles.update(
  customer,
  roleId,
  isSystemRole = NULL,
  etag = NULL,
  kind = NULL,
  roleName = NULL,
  rolePrivileges = NULL,
  isSuperAdminRole = NULL,
  roleDescription = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.roles.update"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customer}{Immutable ID of the G Suite account.}

\item{roleId}{Immutable ID of the role.}

\item{isSystemRole}{Returns \code{true} if this is a pre-defined system role.}

\item{etag}{ETag of the resource.}

\item{kind}{The type of the API resource. This is always \code{admin#directory#role}.}

\item{roleName}{Name of the role.}

\item{rolePrivileges}{The set of privileges that are granted to this role.}

\item{isSuperAdminRole}{Returns \code{true} if the role is a super admin role.}

\item{roleDescription}{A short description of the role.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.rolemanagement
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
