% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemas.patch}
\alias{schemas.patch}
\title{Patch Schema via Apiary Patch Orchestration}
\usage{
schemas.patch(
  customerId,
  schemaKey,
  schemaId = NULL,
  displayName = NULL,
  etag = NULL,
  schemaName = NULL,
  kind = NULL,
  fields = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.schemas.patch"]]$scopes),
  return_response = F
)
}
\arguments{
\item{customerId}{Immutable ID of the G Suite account.}

\item{schemaKey}{Name or immutable ID of the schema.}

\item{schemaId}{The unique identifier of the schema (Read-only)}

\item{displayName}{Display name for the schema.}

\item{etag}{The ETag of the resource.}

\item{schemaName}{The schema's name.}

\item{kind}{Kind of resource this is.}

\item{fields}{A list of fields in the schema.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.userschema
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
