% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users.get}
\alias{users.get}
\title{Retrieves a user.}
\usage{
users.get(
  userKey,
  projection = NULL,
  viewType = NULL,
  customFieldMask = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.users.get"]]$scopes),
  return_response = F
)
}
\arguments{
\item{userKey}{Identifies the user in the API request. The value can be the user's primary email address, alias email address, or unique user ID.}

\item{projection}{What subset of fields to fetch for this user.}

\item{viewType}{Whether to fetch the administrator-only or domain-wide public view of the user. For more information, see \href{/admin-sdk/directory/v1/guides/manage-users#retrieve_users_non_admin}{Retrieve a user as a non-administrator}.}

\item{customFieldMask}{A comma-separated list of schema names. All fields from these schemas are fetched. This should only be set when \code{projection=custom}.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.user
\item https://www.googleapis.com/auth/admin.directory.user.readonly
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
