% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users.insert}
\alias{users.insert}
\title{Creates a user.}
\usage{
users.insert(
  organizations = NULL,
  id = NULL,
  notes = NULL,
  addresses = NULL,
  sshPublicKeys = NULL,
  externalIds = NULL,
  name = NULL,
  ims = NULL,
  orgUnitPath = NULL,
  locations = NULL,
  archived = NULL,
  keywords = NULL,
  password = NULL,
  changePasswordAtNextLogin = NULL,
  recoveryPhone = NULL,
  ipWhitelisted = NULL,
  posixAccounts = NULL,
  includeInGlobalAddressList = NULL,
  gender = NULL,
  languages = NULL,
  phones = NULL,
  relations = NULL,
  recoveryEmail = NULL,
  customSchemas = NULL,
  hashFunction = NULL,
  websites = NULL,
  suspended = NULL,
  primaryEmail = NULL,
  emails = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.users.insert"]]$scopes),
  return_response = F
)
}
\arguments{
\item{organizations}{A list of organizations the user belongs to. The maximum allowed data size for this field is 10Kb.}

\item{id}{The unique ID for the user. A user \code{id} can be used as a user request URI's \code{userKey}.}

\item{notes}{Notes for the user.}

\item{addresses}{A list of the user's addresses. The maximum allowed data size for this field is 10Kb.}

\item{sshPublicKeys}{A list of SSH public keys.}

\item{externalIds}{A list of external IDs for the user, such as an employee or network ID. The maximum allowed data size for this field is 2Kb.}

\item{name}{Holds the given and family names of the user, and the read-only \code{fullName} value. The maximum number of characters in the \code{givenName} and in the \code{familyName} values is 60. In addition, name values support unicode/UTF-8 characters, and can contain spaces, letters (a-z), numbers (0-9), dashes (-), forward slashes (/), and periods (.). For more information about character usage rules, see the \href{http://support.google.com/a/bin/answer.py?answer=33386}{administration help center}. Maximum allowed data size for this field is 1Kb.}

\item{ims}{The user's Instant Messenger (IM) accounts. A user account can have multiple ims properties. But, only one of these ims properties can be the primary IM contact. The maximum allowed data size for this field is 2Kb.}

\item{orgUnitPath}{The full path of the parent organization associated with the user. If the parent organization is the top-level, it is represented as a forward slash (\code{/}).}

\item{locations}{The user's locations. The maximum allowed data size for this field is 10Kb.}

\item{archived}{Indicates if user is archived.}

\item{keywords}{The user's keywords. The maximum allowed data size for this field is 1Kb.}

\item{password}{User's password}

\item{changePasswordAtNextLogin}{Indicates if the user is forced to change their password at next login. This setting doesn't apply when \href{https://support.google.com/a/answer/60224}{the user signs in via a third-party identity provider}.}

\item{recoveryPhone}{Recovery phone of the user. The phone number must be in the E.164 format, starting with the plus sign (+). Example: \emph{+16506661212}.}

\item{ipWhitelisted}{If \code{true}, the user's IP address is \href{http://support.google.com/a/bin/answer.py?answer=60752}{white listed}.}

\item{posixAccounts}{A list of \href{http://www.opengroup.org/austin/papers/posix_faq.html}{POSIX} account information for the user.}

\item{includeInGlobalAddressList}{Indicates if the user's profile is visible in the G Suite global address list when the contact sharing feature is enabled for the domain. For more information about excluding user profiles, see the \href{http://support.google.com/a/bin/answer.py?answer=1285988}{administration help center}.}

\item{gender}{The user's gender. The maximum allowed data size for this field is 1Kb.}

\item{languages}{The user's languages. The maximum allowed data size for this field is 1Kb.}

\item{phones}{A list of the user's phone numbers. The maximum allowed data size for this field is 1Kb.}

\item{relations}{A list of the user's relationships to other users. The maximum allowed data size for this field is 2Kb.}

\item{recoveryEmail}{Recovery email of the user.}

\item{customSchemas}{Custom fields of the user.}

\item{hashFunction}{Stores the hash format of the password property. We recommend sending the \code{password} property value as a base 16 bit hexadecimal-encoded hash value. Set the \code{hashFunction} values as either the \href{http://wikipedia.org/wiki/SHA-1}{SHA-1}, \href{http://wikipedia.org/wiki/MD5}{MD5}, or \href{https://en.wikipedia.org/wiki/Crypt_(C)}{crypt} hash format.}

\item{websites}{The user's websites. The maximum allowed data size for this field is 2Kb.}

\item{suspended}{Indicates if user is suspended.}

\item{primaryEmail}{The user's primary email address. This property is required in a request to create a user account. The \code{primaryEmail} must be unique and cannot be an alias of another user.}

\item{emails}{A list of the user's email addresses. The maximum allowed data size for this field is 10Kb.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.user
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
