% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users.list}
\alias{users.list}
\title{Retrieves a paginated list of either deleted users or all users in a domain.}
\usage{
users.list(
  showDeleted = NULL,
  domain = NULL,
  query = NULL,
  orderBy = NULL,
  pageToken = NULL,
  sortOrder = NULL,
  maxResults = NULL,
  projection = NULL,
  customFieldMask = NULL,
  viewType = NULL,
  customer = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.users.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{showDeleted}{If set to \code{true}, retrieves the list of deleted users. (Default: \code{false})}

\item{domain}{The domain name. Use this field to get fields from only one domain. To return all domains for a customer account, use the \code{customer} query parameter instead. Either the \code{customer} or the \code{domain} parameter must be provided.}

\item{query}{Query string for searching user fields. For more information on constructing user queries, see \href{/admin-sdk/directory/v1/guides/search-users}{Search for Users}.}

\item{orderBy}{Property to use for sorting results.}

\item{pageToken}{Token to specify next page in the list}

\item{sortOrder}{Whether to return results in ascending or descending order.}

\item{maxResults}{Maximum number of results to return.}

\item{projection}{What subset of fields to fetch for this user.}

\item{customFieldMask}{A comma-separated list of schema names. All fields from these schemas are fetched. This should only be set when \code{projection=custom}.}

\item{viewType}{Whether to fetch the administrator-only or domain-wide public view of the user. For more information, see \href{/admin-sdk/directory/v1/guides/manage-users#retrieve_users_non_admin}{Retrieve a user as a non-administrator}.}

\item{customer}{The unique ID for the customer's G Suite account. In case of a multi-domain account, to fetch all groups for a customer, fill this field instead of domain. You can also use the \code{my_customer} alias to represent your account's \code{customerId}. The \code{customerId} is also returned as part of the \href{/admin-sdk/directory/v1/reference/users}{Users resource}. Either the \code{customer} or the \code{domain} parameter must be provided.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.user
\item https://www.googleapis.com/auth/admin.directory.user.readonly
\item https://www.googleapis.com/auth/cloud-platform
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
