% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users.makeAdmin}
\alias{users.makeAdmin}
\title{Makes a user a super administrator.}
\usage{
users.makeAdmin(
  userKey,
  status = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.users.makeAdmin"]]$scopes),
  return_response = F
)
}
\arguments{
\item{userKey}{Identifies the user in the API request. The value can be the user's primary email address, alias email address, or unique user ID.}

\item{status}{Indicates the administrator status of the user.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.user
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
