% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users.undelete}
\alias{users.undelete}
\title{Undeletes a deleted user.}
\usage{
users.undelete(
  userKey,
  orgUnitPath = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.users.undelete"]]$scopes),
  return_response = F
)
}
\arguments{
\item{userKey}{The immutable id of the user}

\item{orgUnitPath}{OrgUnit of User}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.user
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
