% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users.watch}
\alias{users.watch}
\title{Watch for changes in users list}
\usage{
users.watch(
  maxResults = NULL,
  projection = NULL,
  viewType = NULL,
  sortOrder = NULL,
  event = NULL,
  customFieldMask = NULL,
  showDeleted = NULL,
  orderBy = NULL,
  domain = NULL,
  customer = NULL,
  pageToken = NULL,
  query = NULL,
  token = NULL,
  id = NULL,
  expiration = NULL,
  address = NULL,
  params = NULL,
  payload = NULL,
  resourceId = NULL,
  kind = NULL,
  resourceUri = NULL,
  type = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.users.watch"]]$scopes),
  return_response = F
)
}
\arguments{
\item{maxResults}{Maximum number of results to return.}

\item{projection}{What subset of fields to fetch for this user.}

\item{viewType}{Whether to fetch the administrator-only or domain-wide public view of the user. For more information, see \href{/admin-sdk/directory/v1/guides/manage-users#retrieve_users_non_admin}{Retrieve a user as a non-administrator}.}

\item{sortOrder}{Whether to return results in ascending or descending order.}

\item{event}{Events to watch for.}

\item{customFieldMask}{Comma-separated list of schema names. All fields from these schemas are fetched. This should only be set when projection=custom.}

\item{showDeleted}{If set to true, retrieves the list of deleted users. (Default: false)}

\item{orderBy}{Column to use for sorting results}

\item{domain}{Name of the domain. Fill this field to get users from only this domain. To return all users in a multi-domain fill customer field instead."}

\item{customer}{Immutable ID of the G Suite account. In case of multi-domain, to fetch all users for a customer, fill this field instead of domain.}

\item{pageToken}{Token to specify next page in the list}

\item{query}{Query string search. Should be of the form "". Complete documentation is at https: //developers.google.com/admin-sdk/directory/v1/guides/search-users}

\item{token}{An arbitrary string delivered to the target address with each notification delivered over this channel. Optional.}

\item{id}{A UUID or similar unique string that identifies this channel.}

\item{expiration}{Date and time of notification channel expiration, expressed as a Unix timestamp, in milliseconds. Optional.}

\item{address}{The address where notifications are delivered for this channel.}

\item{params}{Additional parameters controlling delivery channel behavior. Optional.}

\item{payload}{A Boolean value to indicate whether payload is wanted. Optional.}

\item{resourceId}{An opaque ID that identifies the resource being watched on this channel. Stable across different API versions.}

\item{kind}{Identifies this as a notification channel used to watch for changes to a resource, which is \code{api#channel}.}

\item{resourceUri}{A version-specific identifier for the watched resource.}

\item{type}{The type of delivery mechanism used for this channel.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.user
\item https://www.googleapis.com/auth/admin.directory.user.readonly
\item https://www.googleapis.com/auth/cloud-platform
}
}
\seealso{
\href{http://developers.google.com/admin-sdk/}{Google Documentation}
}
